User interface, extraction from firebase and add user to waitlist works
