+++ app/src/main/java/com/example/pixel_events/MainActivity.java
@@ -12,6 +12,7 @@ import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
 import com.example.pixel_events.database.DatabaseHandler;
+import com.example.pixel_events.entrant.EventDetailsLauncherActivity;
 import com.example.pixel_events.events.EventActivity;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.firestore.FirebaseFirestore;
@@ -51,7 +52,7 @@ public class MainActivity extends AppCompatActivity {
         addFormButton = findViewById(R.id.addEvent);
 
         addFormButton.setOnClickListener(v -> {
-            Intent intent = new Intent(MainActivity.this, EventActivity.class);
+            Intent intent = new Intent(MainActivity.this, EventDetailsLauncherActivity.class);
             startActivity(intent);
         });
     }
